Resubmitting this because it was easier for me  to do while I’m still learning more than the basics of github forks and pulls.
Text suggestions:

The research page doesn’t actually show the avatars of the team members
so I removed the mention of that in the project builder research page
to make the help text accurate

I added a line to clarify where the user subject count comes from on
the subject uploader page. Partially addresses #1929

Also I found a few things out of date or confusing on the Zooniverse
lab policies. Like you do get formal feedback in terms of  a google
form that is sent at beta, but the policies only mention you should
check Talk. Fixed a typo where there was an errant ‘_’ and clarified
where the buttons for submitting for review and launch are located.
Changed discussion pages to say ‘Talk’ since that’s what they are
referred to elsewhere.