Clean up display of owned cards (collections, favourites etc)
When project cards and owned cards were separated, the CSS was not cleanly split. Owned cards are wrongly stretched vertically to 175px. This change separates out the CSS classes to avoid accidental collisions like this in future.

Also adding some space below collection titles as the text was too close to the button.