downcase the branch name on stage deploy
the domain name of the staged URL is case insensitive and browsers will downcast them before requesting the URL. However our static proxy takes the subdomain and uses it as an s3 bucket path which may not match the branches actual case. This PR will force the bucket path to lowercase to match the domain case via the browser.