added an explicit help line about how to swap subject sets
It still wasn’t clear to me that I didn’t need to bring the site back
to development to swap subject sets. I think an explicit line to remind
project owners/collaborators especially when they haven’t been to the
project builder in a while might be helpful. Feel free to reject if you
think the text on the visibility page is already sufficient