help text addition to visability page  (#2889)
* add in help line for results page

add in help line for results page

* fixed the accidental R deletion

fixed the accidental R deletion

* added in help description for how to reorder workflows

added in a line of text to the workflow page help to tell project
builders how to reorder the listing of their workflows on the project
landing page.

* add in a description of ! in header of manifest file

! in the field header then the values will be shown in Talk only and
not in the classification interface. Adding in to the subject set help
info about this for project owners.

* added an explicit help line about how to swap subject sets

It still wasn’t clear to me that I didn’t need to bring the site back
to development to swap subject sets. I think an explicit line to remind
project owners/collaborators especially when they haven’t been to the
project builder in a while might be helpful. Feel free to reject if you
think the text on the visibility page is already sufficient