Adding in line about how to get your allowance uploaded (#3460)
* Adding in line about how to get your allowance uploaded

Adding in a line about how to contact the Zooniverse team to get your subject upload allowance increased. I had gotten an email from an astronomer interested in using the Zooniverse project builder, but didn't know how to get the allowance increased, so adding a line here like what is on the policies page asking the project owner to contact the Zoonvierse team so it's clear. Not sure if I did the href correctly.

* shorter sentence

shorter sentence

* more edits

shorted sentence

* fixed " to '

* ensure the link renders as html