Audio with image subject (#3878)
Added a red vertical line at the x-position in the direct image, with a
position that corresponds to the current audio track position. Tested
in Chrome and Safari.

Attempt to respect the recommended React rendering flow, whereby
updates are passed as props to child nodes.

Both components are functional and position correctly in Safari and
Chrome. The Safari audio player does not have a position slider, so the
audio track position cannot be changed rapidly except in 30 s jumps.