Autofocus question tasks (#3970)
* Add autofocus prop to GenericTask
Adds the autoFocus prop, which focuses the task container div if true.

* Autofocus single and multiple answer questions
Single answer questions autofocus any selected answer, otherwise autofocus the question.

* Autofocus first selected multiple choice answer
Autofocus the first multiple choice answer, if it exists. Otherwise, autofocus the question.

* Add componentWillReceiveProps for back-to-back questions

* Fix generic task linter warnings

* Tutorials remember document focus
Store the active element when tutorials open, and restore focus when the tutorial closes.

* Autofocus the generic task container on prop changes

Checks if it should autofocus when the question prop changes.
Replaces componentWillReceiveProps with componentDidUpdate so that we autofocus after a render, not before.
Rewrites componentDidMount for consistency with componentDidUpdate.