Fixed options autosave and propagation
hyphen can't exist in key in the generic editor render, no idea why, but throws an error
while converting it to json. Also fixed propagating options to the tool. Will only
add the gridOption variables to toolProps if the tool is the pointGrid tool to avoid
spamming values when it's not necessary.