Fix drag handels when rotated
This change makes it so the corner diagnal to the moving handel stays
fixed durring the resize.  This makes for a much more natural editor.

In addition the rotation point was changed from the center of the
rectangle to the upper left corner (the same point as the `x, y` point
of the rectangle).  This brings it in line with other graphing programs
way of defining a rotated rectange (e.g. matplotlib).