Store workflow task history in Classifier state. (#4450)
* Store current task key in classifier state
Get the current task, and corresponding annotation, from classifier state.
Add a simple handler for task navigation changes.

* Split onNewTask into onNextTask and onPrevTask

* Check for feedback on Next and Done
Check for feedback on next task and on complete.
Hold off on completing the classification until feedback is complete.
If we're showing a summary, set the task key to null, which is a bit of a hack but stops the feedback modal from appearing after the summary step.

* Remove duplicate code from completeClassification

* track workflowTaskHistory in Classifier state
Save task history, rather than just the current task key. This allows us to start to decouple workflow navigation history from the annotations array.

* Hijack the Done & Talk link
Prevent the default event on the Done & Talk link. Instead, show feedback, then submit the classification and navigate to the Talk page.
NB. this disables cmd-click to open the link in a new tab.

* Ignore cmd-click on the Done & Talk link

* Use path, not URL, for subject Talk page