Grid draw tool (#4413)
* Added grid drawing tool

* Added x and y offset to tool options

* Changed tool name

* Fixed options autosave and propagation

hyphen can't exist in key in the generic editor render, no idea why, but throws an error
while converting it to json. Also fixed propagating options to the tool. Will only
add the gridOption variables to toolProps if the tool is the pointGrid tool to avoid
spamming values when it's not necessary.

* Added opacity and color settings

* Converted coffee-react to ES6

* Cleanup

* Cleanup

* Checking to ensure mark is inside grid

* Converted createReactClass to class extends React.Component

* Added pointGrid to dev classifier

* Fixed ESLint errors

* Added icon for point grid

* Cleanup module export

* Fixed delete button positioning

Delete button would be rendered off the screen in certain situations.
This moves the delete button depending on if the grid item clicked
is in the top-most row or right-most column.