Classifier: Don't update classification prop directly (#4484)
* Update annotations instead of classification
Replace props.classification.annotations with props.annotations throughout the classifier. Pass a copy of the annotations array down as a prop. Update that array in classifier state rather than modifying the classification directly.
Only update the classification once the workflow tasks are finished for a subject.

* Fix the dev classifier
Allow the classifier to pass annotations to its children. Displays these in the dev classifier.

* Remember my annotations if I leave the classifier
Classifications persist across all project pages, so store annotations there if we leave the classifier and come back.

* Classifier: remove componentDidUpdate

* Don't clear annotations after the last task
This can interfere with feedback, and with the summary step.
Instead we initialise annotations in componentWillReceiveProps, when a new classification is received.

* Pass annotations directly to ModelRenderer