Drawing subtasks (#4757)
* Fix subtask form prop
onFormClose, not onDetailsFormClose.

* Add a timeout when text and question tasks mount
Add a slight delay to componentDidMount, so that we wait for the component to be added to the DOM before acting on its DOM nodes. This fixes tasks in dialogs, where componentDidMount fires before the dialog is added to the DOM.

* Tidy up text task resize
Get current row count from component state.
Tidy up new row count for readability.
Add scrollHeight to mock textarea in tests.