Store a single subject queue and classification (#4790)
* Store a single subject queue and classification
Remove upcomingSubjects.forWorkflow and currentClassifications.forWorkflow.
Simplify redux actions to work with a single classification and queue.
Update ProjectClassifyPage to use props.classification and props.upcomingSubjects.

* rework createNewClassification
Remove duplicated code for filtering the subject query by subject set.

* Rename resetClassification
Rename resetClassification to nextSubject.
Create a new classification if there's a new subject after shifting the old one from the subject queue.
Reset the classification and queue when emptying the queue.

* Rework subject queue logic
Validate classifications by checking workflow ID.
Reset the subject queue (and classification) if the current classification is invalid.
Rename createNewClassification to refillSubjectQueue.
Refill the subject queue if it's empty after a props change.
Only resume a classification if the component mounts with a valid classification for the currently loaded workflow.

* Refill the queue before it's empty
Refill the queue when the length is less than two. Otherwise the classifier will unmount and reload at the end of each batch, while waiting for a new subject to classify.

* Add some comments
Add some comments to explain the subject queue logic.
Fix some minor linter warnings.

* add subject queue tests
Test that subjects are fetched when the component mounts with no classification.
Test that the queue is reset when the workflow changes.
Test that the queue is reset if the classification prop is invalid.
Test that the classification is resumed if it matches the workflow.

* Refactor refillSubjectQueue
Move subject set code into the fetchSubjects action producer. Extract awaitSubjects and awaitSubjectQueue into helper functions.
Fix componentDidUpdate so that the queue is only reset once when the workflow doesn't match the current classification.

* Hold off on workflow rendering
Don't render until a workflow has loaded. This should ensure that a workflow is present when the classifier mounts.

* Add current workflow to redux store
Update WorkflowSelection to store the selected workflow in redux state.

* Remove currentWorkflowForProject

* Fix workflow selection
Check for a workflow before rendering children.
Set loadingSelectedWorkflow after updating the stored workflow.