Extract the classifier subject queue to the redux store (#4781)
* Extract subject queue to redux
Extract the subject queue code from the classifier and into redux actions for filling the queues, emptying them and getting the next subject to classify.
Update ProjectClassifyPage.getNextSubject to call the new actions.

* Refactor creating a classification
Rewrite createNewClassification to return a classification instead of a promise to create a classification.
Move classification creation from componentWillReceiveProps to componentDidUpdate.

* Create new classifications on subject change
Remove the next subject promises and create a new classification when a new subject prop is passed in from the store.

* Check index before updating annotations
Check that an annotation exists before updating the annotations array.

* Pass subject prop through to classifier
The classifier should now get the subject from the redux store.

* Move currentClassifications to redux store
Store currentClassifications.forWorkflow and remove global variable.
Move createClassification to the ducks file.
Create a new classification when a new subject is requested.
Remount the classifier on workflow change.

* Lint ProjectClassifyPage

* Remove subjectToLoad
It kept track of the current subject while loading new subjects but no longer seems to be used anywhere.

* Fix classifier annotations bug
Set state before completing a classification, to avoid overwriting the new classification with old annotations.

* Fix event handlers that empty the queue
Empty the stored queue on login/logout or when a subject set is updated.

* Reload subject when resuming a classification
Add a resumeClassification action, which loads the correct subject for an existing classification.
Rename the nextSubject action to createClassification, since it's only called when creating a new classification for a subject.

* Pass subject prop to dev classifier

* Get classification subjects from API if necessary
classification.get('subjects') throws an error, so use apiClient.type('subjects') to get subjects for a classification.

* Remove _subjects & _workflow hacks
Remove _subjects and _workflows from classification. Get linked subjects from the API client instead of using _subjects. _workflows isn't used anywhere in the classifier.

* Remove subject from redux state
Remove the current subject from redux state. Get the current subject from the workflow queue instead, when we need it, and store it in classifier state along with the current classification.
Shift the current subject from the queue at the end of a classification, not the beginning.
Unshift a new subject onto the queue if we have to load one after resuming a classification.
Add a RESUME_CLASSIFICATION action to make debugging the classifier lifecycle a little easier.

* Store a single subject queue and classification (#4790)

* Store a single subject queue and classification
Remove upcomingSubjects.forWorkflow and currentClassifications.forWorkflow.
Simplify redux actions to work with a single classification and queue.
Update ProjectClassifyPage to use props.classification and props.upcomingSubjects.

* rework createNewClassification
Remove duplicated code for filtering the subject query by subject set.

* Rename resetClassification
Rename resetClassification to nextSubject.
Create a new classification if there's a new subject after shifting the old one from the subject queue.
Reset the classification and queue when emptying the queue.

* Rework subject queue logic
Validate classifications by checking workflow ID.
Reset the subject queue (and classification) if the current classification is invalid.
Rename createNewClassification to refillSubjectQueue.
Refill the subject queue if it's empty after a props change.
Only resume a classification if the component mounts with a valid classification for the currently loaded workflow.

* Refill the queue before it's empty
Refill the queue when the length is less than two. Otherwise the classifier will unmount and reload at the end of each batch, while waiting for a new subject to classify.

* Add some comments
Add some comments to explain the subject queue logic.
Fix some minor linter warnings.

* add subject queue tests
Test that subjects are fetched when the component mounts with no classification.
Test that the queue is reset when the workflow changes.
Test that the queue is reset if the classification prop is invalid.
Test that the classification is resumed if it matches the workflow.

* Refactor refillSubjectQueue
Move subject set code into the fetchSubjects action producer. Extract awaitSubjects and awaitSubjectQueue into helper functions.
Fix componentDidUpdate so that the queue is only reset once when the workflow doesn't match the current classification.

* Hold off on workflow rendering
Don't render until a workflow has loaded. This should ensure that a workflow is present when the classifier mounts.

* Add current workflow to redux store
Update WorkflowSelection to store the selected workflow in redux state.

* Remove currentWorkflowForProject

* Fix workflow selection
Check for a workflow before rendering children.
Set loadingSelectedWorkflow after updating the stored workflow.