Planet Four fan tool (#4875)
* Add a simple radial line drawing tool
Create a draggable radial line, with a single drag handle to alter the radius and rotation angle.

* Fan tool: add spread angle
Add mark.spread. Rename mark.angle to mark.rotation.
Use default values for initial mark.
Add a line and drag handles to represent the fan spread.

* Add a fan shape
Add a draggable fan shape path.
Increase the default spread.

* Add a fan tool icon

* Add fan tool to project builder
Add fan as an experimental feature.
Add fan tool to the drawing task editor if the experimental feature is enabled.

* Default fan should point right
Swap the x and y axes of the fan, so that the default shape points to the right

* Store the full spread angle
Use half the spread angle for rendering calculations, but store the full spread angle on the mark.

* Add Fan.getCursorAngle(cursor, mark)
Add a helper to calculate the rotation angle for a given cursor position and mark.

* Test tool static methods
Add tests for mark initial placement, and for mark initial radius and rotation angle.

* Add missing scale prop
Scale is passed down to the DragHandle components.

* Test the rotation drag handle
Test that handleRotate updates the mark rotation and radius correctly.

* Test the spread drag handle
Test the spread drag handler for a variety of rotations and cursor click positions.
Expect the spread angle to be limited to 180 degrees maximum.

* Add more test rotation angle positions

* Limit spread angles to 180 degrees
Handle large spread angles correctly.
Limit the maximum spread angle to 180 degrees.

* Limit mark rotation angle in the tests
Real angles lie between +180 and -180.

* Simplify large tests
Run one test per iteration for looping tests. This should make it easier to see which iteration broke if a test fails.

* Only use positive angles from 0 to 360
Return a value between 0 and 360 from Fan.getCursorAngle().
Update tests to remove negative angles.

* Lint the tests