Classification queue: save each queued classification once only (#4923)
* Add a test for slow network connections
Stub resource save to return an unresolved Promise, then check that save is only called once for each classification added to the queue.

* Add a failed classification queue
Create an empty array of failed classifications when flushing the queue. Add to that array each time a classification fails, unless the error status is 422.

* Fix async test for updating the queue
Return a promise when flushing the classification queue.
Wait for that promise to resolve, or fail, before testing the queue.

* Remove unused code
Successful saves don't need to update the failed classification queue, so remove that code completely.

* Fix the sent classification test
Wait for add() to resolve before testing the results of a successful save.

* Add ClassificationQueue.store(classification)
Add a store method, which takes a classification and adds it to the saved queue.

* Handle 422 errors
Discard malformed classifications, as they will never be saved by Panoptes.

* Remove FakeLocalStorage
window.localStorage is supported by JSDOM now, so use that in the tests.

* Clear localStorage between tests
Otherwise, each test loads the stored classifications from the previous test.

* Fix FakeResourceType.save
Add a missing return to the save method.

* Catch assertion errors
Catch errors thrown by failing tests.

* Set a default for onClassificationSaved
Make sure the classification save callback is defined before invoking it.

* Reorganise tests
Add some describe and beforeEach blocks.
Break up expectations into separate it blocks, to make them easier to debug when they fail.

* Remove redundant if block
Promise.all will resolve with an empty array if the pending classifications map is empty, so remove the if block that wraps it.

* Fix tests
The second set of tests were failing because apiClient wasn't properly defined.
Fix the done callback for the second set of tests so that it gets invoked properly when the tests fail.