Fix single answer feedback when the answer is 0 (#4983)
* Update single answer feedback tests
Change the correct answer test to test for the case where the answer is 0 (ie. falsy) but correct.
Update the incorrect answer test to make sure we aren't being tripped up by 0 values.

* Handle answers of 0 correctly
Compare with -1 so that 0 returns true, when determining the answer ID.
Rename annotations to answer for clarity.