Connect the dev classifier to the classify store (#4978)
* Use cheap-module-source-map in development
Enable debugging of cross-origin errors in dev mode by changing the source map setting. See https://reactjs.org/docs/cross-origin-errors.html#webpack

* Add error handling to dev classifier
Add a simple error handler, which dumps the error stack to the page if the classifier crashes.

* Store the classification when resuming
Replace the current classification with the supplied classification when resuming a classification.

* Add a workflow prop to the ClassificationViewer
We're getting the classification from the Redux store, so classification._workflow won't exist any more.

* Connect the dev classifier to the classify store
Store the mock workflow and subjects.
Create a new classification rather than using the mock data.

* Fix the Next button
Refresh the subject queue and load the next subject when the Next button is clicked.

* Initialise on mount
Build a subject queue and new classification on mount

* Tidy up the render method
Replace ternary expressions with if statements.
Add a loading message while the classifier initialises.

* Linting syntax

* Fix dev classifier for anonymous users
Build a new subject queue and classification if an update reset the classification to null.