Add Jenkinsfile to automatically stage branches (#4753)
Create the /src directory with the node user. Jenkins launches the container as the jenkins user by default
(https://issues.jenkins-ci.org/browse/JENKINS-38438) and in this case,
that user is specified by the node alpine image. This will allow the
system to create the /src/dist dir as node user:group avoiding
permissions errors.
Add stage-with-jenkins npm script.