Convert Draggable to ES6 (#5011)
* Convert Draggable to ES6
Convert the Draggable component and update all the drawing tools that require it.

* Convert Draggable to functional component

* Add listeners to document.body
Add drag and end listeners to the document body.

* Don't redefine event types on end
Define move and end events once, on drag start, and use those same types when removing the event listeners at the end of a drag.

* Use React.memo
Add Rect.memo so that Draggable only runs once (for a given set of props) to render children.

* Fix FrameAnnotator tests
Replace mount() with shallow
Remove tests that were actually testing the SVG components.

* Test SVG renderer props
Move tests for the SVG component props out of the FrameAnnotator tests and run them against the SVG component instead.

* Rewrite ComboTask.getSVGProps()
Rewrite getSVGProps to merge the return values of TaskComponent.getSVGProps() for each of the combo's child tasks, then return the merged value for all of those tasks.

* Fix FrameViewer tests
React.memo throws an unknown tag error in enzyme, so use shallow rendering to avoid full rendering of draggable components.

* Add Draggable tests
Add tests for drag start and drag end, from a mouse and from touch events.

* Test callbacks
Add tests for the onStart, onDrag and onEnd callback props.

* Test the disabled component
Make sure it does nothing when disabled.

* Add pointer event support
Add a pointerdown handler and tests for pointer events.

* Add onPointerDown to drawing root
This allows drawing marks to be selected when the Draggable component is dispatching pointer events.

* Test event.preventDefault()
Test that the default event is canceled when a drag interaction starts. I think this allows draggable elements to be nested without interfering with each other.

* Rename coordinates object
Rename the object that's used to pass pageX and pageY to callbacks, to avoid confusion with the original event.

* Remove repeated code from the tests
Move the drag interaction tests into a testDrag() function that can be called with mouse, touch or pointer events.