Test feedback helpers (#5068)
* Test feedback helpers
Add some quick tests for the shared feedback helper functions. Test that rules are correctly extracted from workflow tasks, and that gold standard subject rules are correctly merged with workflow rules when feedback is enabled. Test that feedback rules are copied when rule IDs are specified as numbers and when they are strings.

* Allow numerical values in rules
Check rule values with _.isFinite to allow numbers but fail NaN, Infinity etc.

* Compare rule IDs as strings
Compare subject rule IDs and workflow rule IDs as strings when matching subject feedback to task feedback.