Add a load workflows action to the redux classify store (#5124)
* Add a loadWorkflows action creator
Add an action creator to load a workflow, given a workflow ID, locale and optional user project preferences. It should clear out the existing workflow and subject queue, load the workflow and its translations strings and (optionally) update user project preferences with the workflow ID.
Adds tests for the new action and makes some changes to the translations load action to make it easier to test.

* Replace workflow loading code
Replace workflow loading code with calls to the redux loadWorkflows action in the project home workflow buttons and the workflow assignment dialog.

* Clear preferences on error
If loading a workflow errors, clear the workflow ID from user preferences.

* Use Promise.all()
Use Promise.all() to set the workflow when both workflows and translations have been received from Panoptes.
Update promise chain to use arrow functions.

* Return store state from fakeDispatch
Set up a mock store and return the store state from fakeDispatch. Update the tests to inspect the state that's modified by the dispatcher.
Add tests for clearing the workflow, emptying the subject queue and test that state is not mutated by any of the dispatched actions.

* Add RESET action
Add a reset action, which returns the classify store to its initial state.

* Reset store on new workflow
Reset the classify store to its initial state before loading a new workflow.

* Reset classification state on workflow selection
Call the reset() action before loading a new workflow during workflow selection.

* Check for workflow when refilling subject queue
Avoid a bug where the classify page would try to refill the queue while the workflow is still null.

* Clear stored preferences on 404 errors
Check for 404 errors when retrieving workflows and clear the stored workflow ID from all user preferences if it doesn't exist for that user.
Don't clear the project set workflow for successful requests, as it turns out it's used to unlock workflows on the project home page.

* Use a button for HomeWorkflowButton
Replace the link used to select project workflows with a button.
Replace the span for disabled links with a disabled button.
Update workflow button tests.

* Use Promise.all in workflow selection
Use Promise.all to request workflows and translations during workflow selection, so that we don't update workflow state until we have both the workflow and its translation strings.

* Raise an error for empty API responses
Raise an error if Panoptes returns an empty response for the workflow.
Lint a couple of small syntax errors.

* Simplify the loadWorkflows promise
Handle undefined workflow errors in the API request function.
Simplify the promise chain so that it generates a single action, then dispatches it.

* spec error behaviour for loadWorkflows
Add tests for API error cases. User preferences should be cleared if the workflow does not exist, but should save the workflow ID otherwise so that we can try to recover by reloading the page.