Adds German translation to PFE
Provides a German translation for major parts of Panoptes Front-End. Missing texts will be provided in a second commit in a few days (FAQ, Education in the Zooniverse, How to volunteer). I will also hopefully catch any typos in this second commit.

Note: lines 266 – 287 (about publications) are part of the English en.js version, but not in the de.js file that I translated. I added these and translated them, but did not compare both documents furthermore. If there are other differences that need to be taken into account for the translation, please let me know.