Register new locales in one place (#5493)
* Register new locales in one place
Add a locales object which stores all registered locales, then dynamically load this into counterpart. Contributors then don't have to remember to edit the code in two places in order to add a new language.

* Load locales when the app first loads
Move locales out of the project pages and into the top level app code.

* Check for a language in the URL
Check for a language in the URL when the app first loads.