Improve tutorial accessibility (#5535)
* Improve tutorial accessibility
Render tutorial as a section, not a form.
Add a live region to tutorial step content.
Add more descriptive labels to the step controls.

* Name the tutorial step controls

* Remove Dialog.alert()
Remove Dialog.alert(), which only works for form dialogs.
Add an onComplete callback which is called by the Let's Go! button.
Wrap the tutorial in ModalFocus to prevent keyboard access to the classifier and task controls.

* Add role='dialog'
Add a dialog role so that the dialog (and content?) is announced when opened.