Add a featured project editor to the Projects page (#5667)
* Add a featured project editor
 - Adds a FeaturedProjectEditor, which wraps a project card and allows admins to edit it with the project search selector.
 - adds an optional onChange prop to the project search selector, so that we can override the default link navigation behaviour.

* Restrict search to launch approved projects
Add a launchApproved prop to the projects search and restrict featured projects to launch_approved: true.

* Bugfix: update projects in state
Fix a bug where the featured project is not cleared if we change the project in state but continue to update the project prop.

* Styling fix
Range the editing controls to the right.

* Add options prop for project search
Add an options prop to make project search more flexible. Restrict featured projects search to projects which are launch approved and not already featured.

* Fix project admin search
Always redirect to the project's admin page.
Refactor ProjectListItem function as a component.

* Catch null options onChange
The search selector sends a null option on cancel, so catch that in the onChange handler and only fire onChange if we selected a project.