Change "Annoteren" to "Classificeren"
For a project in the Netherlands, we are working on a project for school kids. We would like to have the word 'Annotate' be translated into 'Classificeren' instead of 'Annoteren'. The latter is a slightly archaic word hardly used in Dutch, while the former is used much more (and children will recognize it).