Upgrade panoptes-client to 3.3.2 (#5939)
This fixes a Sugar problem that was introduced in 3.3.0, and installs a security patch for lodash. The window for refreshing expired tokens is increased to 5 minutes, which might help resolve issues where classifications from logged-in volunteers are being recorded as anonymous. 3.3 also included changes to store the auth bearer token more securely in the client.