Delete index.html and commit_id.txt after upload (#5965)
Upload index.html and commit_id.txt with a cache lifetime of 60s. Delete them after upload then upload the rest of the build with a cache lifetime of 604800s (1 week.)

This should fix any scenarios where index.html could be served out to the CDN with a lifetime of 1 week, before it's overwritten with a cache lifetime of 60s.