correctly setup tests to avoid timer recursion
avoid real API calls by stubbing and thus avoid kicking off multiple timeouts that recurse and cause failures in CI mode.

ideally this is refactored to stub just the API code vs the way it is now but this works and the code is the same as in prod so moving on.