fix: prevent tasks linking to themselves
A second go at fixing this, previously fixed in #5925. Pass the current task key as a prop to `NextTaskSelector`. Filter the list of tasks to exclude that task key, so that a task can't link back to itself.

This has to be edited in the editors for each individual task type, but I think I got them all.